+++ modules/newrelic/settings.go
@@ -12,9 +12,9 @@ const defaultTitle = "NewRelic"
 type Settings struct {
 	common *cfg.Common
 
-	apiKey        string `help:"Your New Relic API token."`
-	applicationID int    `help:"The integer ID of the New Relic application you wish to report on."`
-	deployCount   int    `help:"The number of past deploys to display on screen." optional:"true"`
+	apiKey         string        `help:"Your New Relic API token."`
+	deployCount    int           `help:"The number of past deploys to display on screen." optional:"true"`
+	applicationIDs []interface{} `help:"The integer ID of the New Relic application you wish to report on."`
 }
 
 func NewSettingsFromYAML(name string, ymlConfig *config.Config, globalConfig *config.Config) *Settings {
@@ -22,9 +22,9 @@ func NewSettingsFromYAML(name string, ymlConfig *config.Config, globalConfig *co
 	settings := Settings{
 		common: cfg.NewCommonSettingsFromModule(name, defaultTitle, ymlConfig, globalConfig),
 
-		apiKey:        ymlConfig.UString("apiKey", ymlConfig.UString("apikey", os.Getenv("WTF_NEW_RELIC_API_KEY"))),
-		applicationID: ymlConfig.UInt("applicationID"),
-		deployCount:   ymlConfig.UInt("deployCount", 5),
+		apiKey:         ymlConfig.UString("apiKey", os.Getenv("WTF_NEW_RELIC_API_KEY")),
+		deployCount:    ymlConfig.UInt("deployCount", 5),
+		applicationIDs: ymlConfig.UList("applicationIDs"),
 	}
 
 	return &settings